api/requirements.txt
djangorestframework_jwt==1.11.0
APScheduler==3.6.3
pandas==1.1.4
Django==3.1.1
investpy==1.0
djangorestframework==3.12.2
django-cors-headers==3.6.0
python-dotenv==0.15.0
psycopg2==2.8.6